Fix tests
The restart logic update we pushed earlier to address a bug where the SiteGen data is being cleared at the wrong time of the app's lifecycle fixed the problem in Onboarding but brought up other issues outside of the app in WP Admin.

This logic update will restore the old logic but only process it outside of the Onboarding URLs, so we never clear the SiteGen data within Onboarding.