Improve sitegen logic to prevent accidental regenerations
Given that we don't have a back button on the PreviewsStep, nor do we want users to go back after previews are already generated, the GeneratingStep will now check if the store already has a minimum of 3 homepages, which means a generations already happened. It will then skip the generation logic and move to the previews step again.

Items to consider:
- If the user would like to generate more previews they can use the regenerate action.
- If they want to change the logo, that can be done using the "Select & Customize" action.
- The other reason would be a complete regeneration with a different prompt (during onboarding) and that falls outside of the scope of Onboarding currently and the requirements of this iteration. However, they can use onboarding restart to regenerate different options once they're in the dashboard.